Test for checking background color of delete button passes
A `Button` component does not have a `style` prop. Instead, it has only a `color` prop, which has different appearance on iOS and on Android.

To have a consistent look on both platforms, we have to use a `Pressable` to function as a button and set the color of it instead.